+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter_mod_app/constants/routes.dart';
+
 import 'firebase_options.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
@@ -11,6 +11,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import 'package:flutter_mod_app/constants/theme.dart';
 import 'package:flutter_mod_app/core/navigator/navigator.dart';
+import 'package:flutter_mod_app/constants/routes.dart';
 import 'package:flutter_mod_app/core/providers/theme.dart';
 
 void main() async {
