+++ lib/view/screens/auth/authScreen.dart
@@ -2,10 +2,10 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mod_app/constants/appAssets.dart';
 import 'package:flutter_mod_app/constants/routes.dart';
+import 'package:flutter_mod_app/core/controllers/authController.dart';
 import 'package:flutter_mod_app/view/components/bigButton/button.dart';
 import 'package:flutter_mod_app/view/components/textInput/textInput.dart';
 import 'package:flutter_mod_app/constants/colors.dart';
-import 'package:flutter_mod_app/core/services/auth.dart';
 import 'package:flutter_mod_app/core/models/auth_data.dart';
 
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -18,6 +18,7 @@ class AuthScreen extends StatefulWidget {
 }
 
 class _AuthScreenState extends State<AuthScreen> {
+  final AuthController _authController = AuthController();
   String _username = "";
   String _password = "";
   String _error = "";
@@ -44,7 +45,7 @@ class _AuthScreenState extends State<AuthScreen> {
 
   void _onPressLogin() async {
     try {
-      await authWithEmail(_username, _password);
+      await _authController.signWithMail(_username, _password);
       setState(() => _error = "");
       if (!mounted) return;
       Navigator.pushNamed(context, homeRoute,
